Add support for rotating files
One potential use case for `edit-filenames` is to manually rotate
files.  Previously attempts to swap filenames would fail because the
destination files already exist.

If the destination file already exists, check if it's also expected
to be renamed.  If so, defer the rename to later, renaming to an
intermediate temporary name if necessary.