Make the instructions use a multi-line string literal
Use a multi-line string literal for the instructions instead of a
list of single-line strings.  This makes it easier to visualize and
to edit.

This makes the computation for the first non-instructional line more
complicated, however.  Instead of using just the length of the list,
we also must count the number of embedded newlines in the multi-line
literal.  Note that we must still account for the list length since
`spawneditor.edit_temporary` expects a list of lines and will write
each element with an additional newline.